+++ start.sh
@@ -9,7 +9,7 @@ r=()
 r+=("$(find /etc -type f -name 'aliases')")
 r+=("$(find /etc/postfix -type f -name '*.cf')")
 r+=("$(find /etc/dovecot -type f -name '*.conf*')")
-r+=("$(find /srv -type f)")
+r+=("$(find /etc/cron.d -type f -name 'cleanup-*')")
 
 # Replace environment vars
 for d in "${r[@]}"
@@ -30,17 +30,15 @@ done
 # Update root alias
 newaliases
 
-# Start logging
 service syslog-ng start
 
-# Start Dovecot
+service cron start
+
 service dovecot start
 
 # Grace time to prevent SASL authentication method error
 sleep 5
 
-# Start Postfix
 service postfix start
 
-# Show logs
 tail -f /var/log/mail.log